Added util to combine `WriteResult`s
