+++ python/ray/data/datasource/datasink.py
@@ -1,3 +1,4 @@
+import itertools
 import logging
 from dataclasses import dataclass
 from typing import Generic, Iterable, List, Optional, TypeVar
@@ -25,6 +26,20 @@ class WriteResult(Generic[WriteReturnType]):
     size_bytes: int
     # All returned values of `Datasink.write`.
     write_returns: List[WriteReturnType]
+
+    @classmethod
+    def combine(cls, *wrs: "WriteResult") -> "WriteResult":
+        num_rows = sum(wr.num_rows for wr in wrs)
+        size_bytes = sum(wr.size_bytes for wr in wrs)
+        write_returns = list(
+            itertools.chain(*[wr.write_returns for wr in wrs])
+        )
+
+        return WriteResult(
+            num_rows=num_rows,
+            size_bytes=size_bytes,
+            write_returns=write_returns,
+        )
 
 
 @DeveloperAPI
@@ -162,3 +177,20 @@ class DummyOutputDatasink(Datasink[None]):
 
     def on_write_failed(self, error: Exception) -> None:
         self.num_failed += 1
+
+
+def gen_datasink_write_result(
+    write_result_blocks: List[Block],
+) -> WriteResult:
+    import pandas as pd
+
+    assert all(
+        isinstance(block, pd.DataFrame) and len(block) == 1
+        for block in write_result_blocks
+    )
+
+    total_num_rows = sum(result["num_rows"].sum() for result in write_result_blocks)
+    total_size_bytes = sum(result["size_bytes"].sum() for result in write_result_blocks)
+
+    write_returns = [result["write_return"][0] for result in write_result_blocks]
+    return WriteResult(total_num_rows, total_size_bytes, write_returns)
